# Optimize Histogram::clone and in some cases Histogram::add