Optimize Histogram::clone by using a simple derived implementation

Benchmarks for cloning an Histogram created with `Histogram::<u32>::new_with_bounds(1, 100_000, 3)`

Before
`test clone ... bench:       6,513 ns/iter (+/- 66)`
After
`test clone ... bench:         532 ns/iter (+/- 16)`

and also
Optimize Histogram::add when the histogram parameters are equivalent and the target histogram is empty

Before
`test new_from_then_add ... bench:       6,513 ns/iter (+/- 66)`
After
`test new_from_then_add ... bench:         788 ns/iter (+/- 46)`